fix(xhr): do not use private fields to allow compatibility with MSW
Often applications use tools like [MSW](https://mswjs.io/) for testing,
those tools use Proxies over the real implementation.
But because of https://github.com/tc39/proposal-class-fields/issues/106
this is not possible for private fields used in setters when called by a
Proxy.

So a solution is to just use Typescript visibility modifiers.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>